healthid/cli/requirements.txt
python-jose==3.2.0
cryptography==3.0
ecdsa==0.14.1
requests==2.24.0
